requirements.txt
boto3==1.24.19
botocore==1.27.19
ffmpeg-python==0.2.0
matplotlib==3.5.2
numpy==1.21.6
pandas==1.4.3
s3transfer==0.6.0
urllib3==1.26.9
requests==2.28.1
